kadai 1st

+++ app/Http/Controllers/TasklistsController.php
@@ -17,13 +17,22 @@ class TasklistsController extends Controller
      */
     public function index()
     {
-        $tasklists = Tasklist::all();
-        
-        return view('tasklists.index',[
-            'tasklists' => $tasklists,
-            ]);
-        //
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasklists' => $tasklists,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -49,15 +58,14 @@ class TasklistsController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            'status' => 'required|max:10',
-        ]);    
-            
-        $tasklist = new Tasklist;
-        $tasklist->status = $request->status;
-        $tasklist->content = $request->content;
-        $tasklist->save();
-        
-        return redirect('/');
+            'content' => 'required|max:191',
+        ]);
+
+        $request->user()->tasklists()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+        return redirect()->back();
         //
     }
 
@@ -70,11 +78,16 @@ class TasklistsController extends Controller
     public function show($id)
     {
         $tasklist = Tasklist::find($id);
+        if (\Auth::id() == $tasklist->user_id) {
         
         return view('tasklists.show',[
             'tasklist' => $tasklist,
             ]);
+        } else {
+            return redirectly('/');
         //
+        }
+        
     }
 
     /**
@@ -86,11 +99,14 @@ class TasklistsController extends Controller
     public function edit($id)
     {
         $tasklist = Tasklist::find($id);
-         
+        if (\Auth::id() == $tasklist->user_id) {
+
         return view('tasklists.edit',[
             'tasklist' => $tasklist,
             ]);
-        //
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -103,16 +119,19 @@ class TasklistsController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
-            'status' => 'required|max:10',
+            'status' => 'required|max:30',
+            'content'=> 'required|max:20',
         ]);    
         
         $tasklist = Tasklist::find($id);
+         if(\Auth::id() == $task->user_id){
         $tasklist->status = $request->status;
         $tasklist->content = $request->content;
         $tasklist->save();
-        
+         } else {
         return redirect('/');
-        //
+    
+        }
     }
 
     /**
@@ -121,12 +140,20 @@ class TasklistsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-        $tasklist = Tasklist::find($id);
-        $tasklist->delete();
-        
-        return redirect('/');
-        //
+     public function destroy($id)
+     
+     
+     
+     
+        {
+        $tasklist = \App\Tasklist::find($id);
+
+        if (\Auth::id() == $tasklist->user_id) {
+            $tasklist->delete();
+        }else{
+
+        return redirect()->back();
+        }
     }
 }
+
